+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 ## 1.0.1 - 2023-01-10
 
-### Added
-- Fix namespace issue
+### Fixed
 
+- Fix namespace issue
 
 ## 1.0.0 - 2023-01-10
 
 ### Added
+
 - Initial release